move to getter setter file
